[0.3.x] fix: MEMORY_BASIC_INFORMATION64::RegionSize is ULONGLONG
git-svn-id: https://pykd.svn.codeplex.com/svn@87630 9b283d60-5439-405e-af05-b73fd8c4d996

+++ pykd/pymemaccess.h
@@ -154,9 +154,8 @@ inline bool compareMemory( kdlib::MEMOFFSET_64 addr1, kdlib::MEMOFFSET_64 addr2,
 
 inline python::tuple findMemoryRegion( kdlib::MEMOFFSET_64 offset )
 {
-
-    kdlib::MEMOFFSET_64   regionOffset;
-    size_t  regionLength;
+    kdlib::MEMOFFSET_64 regionOffset;
+    unsigned long long  regionLength;
 
     AutoRestorePyState  pystate;
